Commit of project outline:
React + Vite frontend with an Express backend.
Currently loads a simple JSON from the backend and sends it to the front.

Stage 2 is writing a backend API that takes a YouTube playlist URL.
In response, it will send a list of playlists that have been sorted into bins.

Stage 3 is be creating a component to play the playlists.

Stage 4 is hooking the player up to a timer to automatically start/end sessions.

Stage 5 is writing CSS and three.js assets to make it fun to look at.

Stage 6 is adding some developer info.